test: improved coverage
