+++ packages/@postdfm/dfm2ast/__test__/broken/form.dfm
@@ -1,0 +1,4 @@
+object Form1: TForm1
+  Left = 192
+  Font.Name = 'MS Sans Serif'
+  Font.Style = []
